+++ src/main/java/io/apimap/cli/commands/DeleteCommand.java
@@ -20,12 +20,15 @@ under the License.
 package io.apimap.cli.commands;
 
 import io.apimap.api.rest.jsonapi.JsonApiRestResponseWrapper;
+import io.apimap.cli.utils.ConfigurationFileUtil;
 import io.apimap.cli.utils.MetadataUtil;
 import io.apimap.client.IRestClient;
 import io.apimap.client.RestClientConfiguration;
 import io.apimap.client.exception.IncorrectTokenException;
 import io.apimap.file.metadata.MetadataFile;
 import picocli.CommandLine;
+
+import java.io.IOException;
 
 @CommandLine.Command(
         name = "delete",
@@ -35,7 +38,7 @@ import picocli.CommandLine;
 public class DeleteCommand extends ApiCommand implements Runnable {
     @CommandLine.Option(
             names = {"--metadata"},
-            description = "File path to the metadata file to be published. E.g my-api/metadata.apicatalog"
+            description = "File path to the metadata file to be published. E.g my-api/metadata.apimap"
     )
     protected String metadataFilePath;
 
@@ -48,7 +51,7 @@ public class DeleteCommand extends ApiCommand implements Runnable {
 
     @CommandLine.Option(
             names = {"--confirmation"},
-            description = "This will permanently REMOVE ALL information?",
+            description = "This will permanently REMOVE ALL information.",
             interactive = true,
             arity="0..1"
     )
@@ -57,8 +60,15 @@ public class DeleteCommand extends ApiCommand implements Runnable {
     @Override
     public void run() {
         if (this.endpointUrl == null) {
-            System.err.println("[Error] Missing endpoint url");
-            return;
+            try {
+                this.endpointUrl = new ConfigurationFileUtil(ConfigurationFileUtil.FILENAME).readEndpoint();
+            } catch (IOException ignored) {
+            }
+
+            if (this.endpointUrl == null) {
+                System.err.println("[Error] Missing endpoint url");
+                return;
+            }
         }
 
         if (this.metadataFilePath == null) {
@@ -67,7 +77,7 @@ public class DeleteCommand extends ApiCommand implements Runnable {
         }
 
         if (confirmation == null) {
-            String s = System.console().readLine("Continue this dangerous action? y/n: ");
+            String s = System.console().readLine("This will permanently REMOVE ALL information. Continue? y/n: ");
             confirmation = Boolean.valueOf(s) || "y".equalsIgnoreCase(s);
         }