+++ CHANGELOG.md
@@ -2,7 +2,11 @@ Changelog
 ===
 
 ### 21.04.2022
-- asd
+- Added action confirm dialogs on dangerous actions
+- Fixed token file bug
+- Updated documentation
+- Added static endpoint configuration if only one apimap instance is to be used
+- Simplified token list
 
 ### 05.11.2021 
 - First public release